requirements.txt
# Flask Framework et Extensions
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Google Cloud Services
google-cloud-firestore==2.13.1
google-generativeai==0.3.2

# Text-to-Speech
gTTS==2.5.4

# Utilitaires
python-dotenv==1.0.0
requests==2.31.0

# Retry Logic et Gestion d'Erreurs
tenacity==8.2.3

# Tests (optionnel, pour développement)
# pytest==7.4.3
# pytest-cov==4.1.0
# pytest-flask==1.3.0

# Linting et Formatage (optionnel, pour développement)
# black==23.12.0
# flake8==6.1.0
# pylint==3.0.3

# Type Hints (optionnel, pour développement)
# mypy==1.7.1

# Documentation (optionnel)
# sphinx==7.2.6
# flasgger==0.9.7.1  # Pour Swagger/OpenAPI

# Production (décommenter pour déploiement)
gunicorn==21.2.0
gevent==23.9.1